chore(src): refactoring codebase
