+++ netflix/package.json
@@ -43,12 +43,14 @@
     ]
   },
   "devDependencies": {
+    "@types/node-sass": "^4.11.3",
     "@typescript-eslint/eslint-plugin": "^5.45.1",
     "@typescript-eslint/parser": "^5.45.1",
     "eslint": "^8.29.0",
     "eslint-config-xo": "^0.43.1",
     "eslint-config-xo-typescript": "^0.55.1",
     "eslint-plugin-react": "^7.31.11",
+    "node-sass": "^7.0.3",
     "typescript": "^4.9.3"
   }
 }
