+++ netflix/src/dummyFilms.ts
@@ -1,0 +1,39 @@
+const films = [
+  {
+    img: "https://picsum.photos/400/400",
+    name: "Pulp",
+    genre: "Action",
+    year: 2005,
+    key: 1,
+  },
+  {
+    img: "https://picsum.photos/400/400",
+    name: "Fiction",
+    genre: "Action",
+    year: 2005,
+    key: 2,
+  },
+  {
+    img: "https://picsum.photos/400/400",
+    name: "Is",
+    genre: "Action",
+    year: 2005,
+    key: 3,
+  },
+  {
+    img: "https://picsum.photos/400/400",
+    name: "Is",
+    genre: "Action",
+    year: 2005,
+    key: 5,
+  },
+  {
+    img: "https://picsum.photos/400/400",
+    name: "Is",
+    genre: "Action",
+    year: 2005,
+    key: 6,
+  },
+];
+
+export default films;
